-  The CameraUtils.java file has been split into several classes
