+++ src/main/java/com/playsi/aero_cam_sync/client/mixins/CameraMixin.java
@@ -2,6 +2,9 @@ package com.playsi.aero_cam_sync.client.mixins;
 
 import com.playsi.aero_cam_sync.client.config.Config;
 import com.playsi.aero_cam_sync.client.debug.DebugRayRenderer;
+import com.playsi.aero_cam_sync.client.utils.CameraController;
+import com.playsi.aero_cam_sync.client.utils.SubLevelTracker;
+import com.playsi.aero_cam_sync.client.utils.SurfaceRaycaster;
 import dev.ryanhcode.sable.companion.math.Pose3dc;
 import dev.ryanhcode.sable.sublevel.ClientSubLevel;
 import net.minecraft.client.Camera;
@@ -14,9 +17,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import static com.playsi.aero_cam_sync.client.utils.CameraUtils.*;
-
-@Mixin(Camera.class)
+@Mixin(value = Camera.class, priority = 10000)
 public abstract class CameraMixin {
 
     @Inject(method = "setup", at = @At("TAIL"))
@@ -26,23 +27,26 @@ public abstract class CameraMixin {
             float partialTick, CallbackInfo ci) {
 
         if (!Config.MOD_ENABLED.get()) return;
-        if (!shouldApplyTilt()) return;
+        if (!CameraController.shouldApplyTilt()) return;
 
         Minecraft mc = Minecraft.getInstance();
         float deltaTime = mc.getTimer().getRealtimeDeltaTicks();
 
-        ClientSubLevel subLevel = getClientSubLevel(mc.player);
+        ClientSubLevel subLevel = SubLevelTracker.getClientSubLevel(mc.player);
 
         DebugRayRenderer.clear();
 
         Vector3f surfaceNormal = null;
-
         if (subLevel != null) {
             Pose3dc pose = subLevel.renderPose(partialTick);
-            surfaceNormal = getSurfaceNormal(subLevel, pose);
+            surfaceNormal = SurfaceRaycaster.getSurfaceNormal(subLevel, pose);
+
+            if (surfaceNormal == null && Config.DROP_CACHE_ON_ALL_MISS.get()) {
+                SubLevelTracker.invalidateCache();
+            }
         }
 
-        updateSmoothedTilt(surfaceNormal, deltaTime, false);
-        applyTiltToCamera((Camera)(Object) this, partialTick);
+        CameraController.updateSmoothedTilt(surfaceNormal, deltaTime, false);
+        CameraController.applyTiltToCamera((Camera)(Object) this, partialTick);
     }
 }
